Don't strip debugging symbols in the regular Debian images
Since #687, the regular Debian images no longer ship DWARF symbols,
which heavily limits the ability to use gdb and lldb for debugging.

Stripping debugging symbols makes perfect sense to reduce the size
of Docker images, but when this optimization was first introduced
in #483, the idea was to restrict its use to the slim and Alpine images.
That was (unintentionally?) changed in #483, and now debugging symbols
are also stripped in the regular Debian images. Undo that change.